cnn_image/requirements.txt
torch
torchvision
Pillow
python-multipart
uvicorn[standard]
python-json-logger
fastapi
